Refactor cartographer view around controller
