--- salt-marcher/src/apps/cartographer/mode-registry/registry.ts
@@ -1,397 +1,0 @@
-// src/apps/cartographer/mode-registry/registry.ts
-// Kernlogik zur Registrierung von Cartographer-Modi.
-import type {
-    CartographerMode,
-    CartographerModeLifecycleContext,
-} from "../presenter";
-
-export type CartographerModeMapInteraction = "none" | "hex-click";
-
-export type CartographerModePersistence = "read-only" | "manual-save";
-
-export type CartographerModeSidebarUsage = "required" | "optional" | "hidden";
-
-export interface CartographerModeCapabilities {
-    readonly mapInteraction: CartographerModeMapInteraction;
-    readonly persistence: CartographerModePersistence;
-    readonly sidebar: CartographerModeSidebarUsage;
-}
-
-export type CartographerModeWithCapabilities<Caps extends CartographerModeCapabilities> = CartographerMode &
-    (Caps["mapInteraction"] extends "hex-click"
-        ? { onHexClick: NonNullable<CartographerMode["onHexClick"]> }
-        : { onHexClick?: never }) &
-    (Caps["persistence"] extends "manual-save"
-        ? { onSave: NonNullable<CartographerMode["onSave"]> }
-        : { onSave?: never });
-
-export interface CartographerModeMetadata<Caps extends CartographerModeCapabilities = CartographerModeCapabilities> {
-    readonly id: string;
-    readonly label: string;
-    readonly summary: string;
-    readonly keywords?: readonly string[];
-    readonly order?: number;
-    readonly source: string;
-    readonly version?: string;
-    readonly capabilities: Caps;
-}
-
-export interface CartographerModeProvider<Caps extends CartographerModeCapabilities = CartographerModeCapabilities> {
-    readonly metadata: CartographerModeMetadata<Caps>;
-    load(): Promise<CartographerModeWithCapabilities<Caps>>;
-}
-
-export const defineCartographerModeProvider = <Caps extends CartographerModeCapabilities>(
-    provider: CartographerModeProvider<Caps>,
-): CartographerModeProvider<Caps> => provider;
-
-export type NormalizedCartographerModeMetadata = CartographerModeMetadata<CartographerModeCapabilities>;
-
-type ProviderEntry = {
-    readonly provider: CartographerModeProvider;
-    readonly metadata: NormalizedCartographerModeMetadata;
-};
-
-type RegisteredProvider = ProviderEntry & {
-    readonly mode: CartographerMode;
-};
-
-export interface CartographerModeRegistryEntry {
-    readonly metadata: NormalizedCartographerModeMetadata;
-    readonly mode: CartographerMode;
-}
-
-export type CartographerModeRegistryEvent =
-    | { readonly type: "initial"; readonly entries: readonly CartographerModeRegistryEntry[] }
-    | {
-          readonly type: "registered";
-          readonly entry: CartographerModeRegistryEntry;
-          readonly index: number;
-          readonly entries: readonly CartographerModeRegistryEntry[];
-      }
-    | {
-          readonly type: "deregistered";
-          readonly id: string;
-          readonly entries: readonly CartographerModeRegistryEntry[];
-      }
-    | { readonly type: "reset"; readonly entries: readonly CartographerModeRegistryEntry[] };
-
-type CartographerModeRegistryListener = (event: CartographerModeRegistryEvent) => void;
-
-const providers = new Map<string, RegisteredProvider>();
-const listeners = new Set<CartographerModeRegistryListener>();
-
-const MAP_INTERACTIONS: readonly CartographerModeMapInteraction[] = ["none", "hex-click"];
-const PERSISTENCE_MODES: readonly CartographerModePersistence[] = ["read-only", "manual-save"];
-const SIDEBAR_USAGES: readonly CartographerModeSidebarUsage[] = ["required", "optional", "hidden"];
-
-const normalizeCapabilities = (
-    providerId: string,
-    capabilities: CartographerModeCapabilities | undefined,
-): CartographerModeCapabilities => {
-    if (!capabilities) {
-        throw new Error(
-            `[cartographer:mode-registry] provider '${providerId}' must declare capabilities metadata`,
-        );
-    }
-
-    const { mapInteraction, persistence, sidebar } = capabilities as CartographerModeCapabilities;
-
-    if (!MAP_INTERACTIONS.includes(mapInteraction)) {
-        throw new Error(
-            `[cartographer:mode-registry] provider '${providerId}' declared invalid mapInteraction capability '${mapInteraction}'`,
-        );
-    }
-
-    if (!PERSISTENCE_MODES.includes(persistence)) {
-        throw new Error(
-            `[cartographer:mode-registry] provider '${providerId}' declared invalid persistence capability '${persistence}'`,
-        );
-    }
-
-    if (!SIDEBAR_USAGES.includes(sidebar)) {
-        throw new Error(
-            `[cartographer:mode-registry] provider '${providerId}' declared invalid sidebar capability '${sidebar}'`,
-        );
-    }
-
-    return Object.freeze({
-        mapInteraction,
-        persistence,
-        sidebar,
-    } satisfies CartographerModeCapabilities);
-};
-
-const cloneMetadata = (metadata: CartographerModeMetadata): NormalizedCartographerModeMetadata => {
-    const keywords = metadata.keywords ? Object.freeze([...metadata.keywords]) : undefined;
-    const normalized: NormalizedCartographerModeMetadata = Object.freeze({
-        ...metadata,
-        keywords,
-        capabilities: normalizeCapabilities(metadata.id, metadata.capabilities),
-    });
-    return normalized;
-};
-
-const normalizeProvider = (provider: CartographerModeProvider): ProviderEntry => {
-    if (!provider?.metadata?.id) {
-        throw new Error("[cartographer:mode-registry] provider metadata requires an id");
-    }
-    if (!provider.metadata.label) {
-        throw new Error(`[` + "cartographer:mode-registry" + `] provider '${provider.metadata.id}' requires a label`);
-    }
-    if (!provider.metadata.summary) {
-        throw new Error(`[` + "cartographer:mode-registry" + `] provider '${provider.metadata.id}' requires a summary`);
-    }
-    if (!provider.metadata.source) {
-        throw new Error(
-            `[` + "cartographer:mode-registry" + `] provider '${provider.metadata.id}' requires a source identifier`,
-        );
-    }
-
-    const metadata = cloneMetadata(provider.metadata);
-
-    return {
-        provider: {
-            ...provider,
-            metadata,
-        },
-        metadata,
-    } satisfies ProviderEntry;
-};
-
-const validateModeCapabilities = (
-    mode: CartographerMode,
-    metadata: NormalizedCartographerModeMetadata,
-): void => {
-    const { capabilities, id } = metadata;
-
-    if (capabilities.mapInteraction === "hex-click" && typeof mode.onHexClick !== "function") {
-        throw new Error(
-            `[cartographer:mode-registry] mode '${id}' declares mapInteraction 'hex-click' but does not implement onHexClick()`,
-        );
-    }
-
-    if (capabilities.persistence === "manual-save" && typeof mode.onSave !== "function") {
-        throw new Error(
-            `[cartographer:mode-registry] mode '${id}' declares persistence 'manual-save' but does not implement onSave()`,
-        );
-    }
-
-    if (capabilities.mapInteraction === "none" && typeof mode.onHexClick === "function") {
-        console.warn(
-            `[cartographer:mode-registry] mode '${id}' provides onHexClick(), but its capabilities declare mapInteraction 'none'`,
-        );
-    }
-};
-
-const createLazyModeWrapper = (entry: ProviderEntry): CartographerMode => {
-    const { metadata, provider } = entry;
-    let cached: CartographerMode | null = null;
-    let loading: Promise<CartographerMode> | null = null;
-
-    const load = async (): Promise<CartographerMode> => {
-        if (cached) return cached;
-        if (!loading) {
-            loading = provider
-                .load()
-                .then((mode) => {
-                    if (!mode) {
-                        throw new Error(
-                            `[cartographer:mode-registry] provider '${metadata.id}' returned an invalid mode instance`,
-                        );
-                    }
-                    if (mode.id && mode.id !== metadata.id) {
-                        console.warn(
-                            `[cartographer:mode-registry] mode id '${mode.id}' does not match provider id '${metadata.id}'`,
-                        );
-                    }
-                    validateModeCapabilities(mode, metadata);
-                    cached = mode;
-                    return mode;
-                })
-                .catch((error) => {
-                    console.error(
-                        `[cartographer:mode-registry] failed to load mode '${metadata.id}' from '${metadata.source}'`,
-                        error,
-                    );
-                    loading = null;
-                    throw error;
-                });
-        }
-        return loading;
-    };
-
-    type ModeMethodKey = {
-        [K in keyof CartographerMode]: CartographerMode[K] extends (...args: any[]) => any ? K : never;
-    }[keyof CartographerMode];
-
-    type ModeMethod<K extends ModeMethodKey> = Extract<CartographerMode[K], (...args: any[]) => any>;
-
-    const invoke = async <K extends ModeMethodKey>(
-        key: K,
-        ...args: Parameters<ModeMethod<K>>
-    ): Promise<Awaited<ReturnType<ModeMethod<K>>>> => {
-        const mode = await load();
-        const method = mode[key] as ModeMethod<K>;
-        return await method.apply(mode, args);
-    };
-
-    const invokeIfLoaded = async <K extends ModeMethodKey>(
-        key: K,
-        ...args: Parameters<ModeMethod<K>>
-    ): Promise<Awaited<ReturnType<ModeMethod<K>>> | undefined> => {
-        if (!cached && !loading) {
-            await load();
-        }
-        const mode = cached;
-        if (!mode) return undefined;
-        const method = mode[key];
-        if (typeof method !== "function") {
-            return undefined;
-        }
-        return await (method as ModeMethod<K>).apply(mode, args);
-    };
-
-    return {
-        id: metadata.id,
-        label: metadata.label,
-        async onEnter(ctx: CartographerModeLifecycleContext) {
-            return await invoke("onEnter", ctx);
-        },
-        async onExit(ctx) {
-            await invokeIfLoaded("onExit", ctx);
-        },
-        async onFileChange(file, handles, ctx) {
-            return await invoke("onFileChange", file, handles, ctx);
-        },
-        async onHexClick(coord, event, ctx) {
-            if (metadata.capabilities.mapInteraction !== "hex-click") {
-                return;
-            }
-            return await invokeIfLoaded("onHexClick", coord, event, ctx);
-        },
-        async onSave(mode, file, ctx) {
-            if (metadata.capabilities.persistence !== "manual-save") {
-                return undefined;
-            }
-            return await invokeIfLoaded("onSave", mode, file, ctx);
-        },
-    } satisfies CartographerMode;
-};
-
-const createRegisteredProvider = (provider: CartographerModeProvider): RegisteredProvider => {
-    const entry = normalizeProvider(provider);
-    return {
-        ...entry,
-        mode: createLazyModeWrapper(entry),
-    } satisfies RegisteredProvider;
-};
-
-const toRegistryEntry = (provider: RegisteredProvider): CartographerModeRegistryEntry => ({
-    metadata: provider.metadata,
-    mode: provider.mode,
-});
-
-const notifyListeners = (event: CartographerModeRegistryEvent): void => {
-    const snapshot = Array.from(listeners);
-    for (const listener of snapshot) {
-        try {
-            listener(event);
-        } catch (error) {
-            console.error("[cartographer:mode-registry] listener failed", error);
-        }
-    }
-};
-
-const orderValue = (metadata: CartographerModeMetadata): number => {
-    if (metadata.order === undefined || Number.isNaN(metadata.order)) return Number.POSITIVE_INFINITY;
-    return metadata.order;
-};
-
-const getSortedProviders = (): RegisteredProvider[] => {
-    return Array.from(providers.values()).sort((a, b) => {
-        const orderDiff = orderValue(a.metadata) - orderValue(b.metadata);
-        if (orderDiff !== 0) return orderDiff;
-        return a.metadata.label.localeCompare(b.metadata.label, undefined, { sensitivity: "base" });
-    });
-};
-
-const getRegistryEntries = (): readonly CartographerModeRegistryEntry[] => {
-    return getSortedProviders().map(toRegistryEntry);
-};
-
-export const registerCartographerModeProvider = (
-    provider: CartographerModeProvider,
-): (() => void) => {
-    const entry = createRegisteredProvider(provider);
-    const existing = providers.get(entry.metadata.id);
-    if (existing) {
-        throw new Error(
-            `[cartographer:mode-registry] provider with id '${entry.metadata.id}' is already registered by '${existing.metadata.source}'`,
-        );
-    }
-    providers.set(entry.metadata.id, entry);
-
-    const entries = getRegistryEntries();
-    const index = entries.findIndex((candidate) => candidate.metadata.id === entry.metadata.id);
-    if (index >= 0) {
-        notifyListeners({
-            type: "registered",
-            entry: entries[index]!,
-            index,
-            entries,
-        });
-    }
-
-    return () => {
-        const current = providers.get(entry.metadata.id);
-        if (current === entry) {
-            providers.delete(entry.metadata.id);
-            const remaining = getRegistryEntries();
-            notifyListeners({
-                type: "deregistered",
-                id: entry.metadata.id,
-                entries: remaining,
-            });
-        }
-    };
-};
-
-export const unregisterCartographerModeProvider = (id: string): boolean => {
-    const existed = providers.delete(id);
-    if (existed) {
-        const remaining = getRegistryEntries();
-        notifyListeners({ type: "deregistered", id, entries: remaining });
-    }
-    return existed;
-};
-
-export const getCartographerModeMetadataSnapshot = (): readonly NormalizedCartographerModeMetadata[] => {
-    return getSortedProviders().map((entry) => entry.metadata);
-};
-
-export const createCartographerModesSnapshot = (): CartographerMode[] => {
-    return getSortedProviders().map((entry) => entry.mode);
-};
-
-export const createCartographerModeRegistrySnapshot = (): readonly CartographerModeRegistryEntry[] => {
-    return getRegistryEntries();
-};
-
-export const clearCartographerModeRegistry = (): void => {
-    if (providers.size === 0) {
-        return;
-    }
-    providers.clear();
-    notifyListeners({ type: "reset", entries: [] });
-};
-
-export const subscribeToCartographerModeRegistry = (
-    listener: CartographerModeRegistryListener,
-): (() => void) => {
-    listeners.add(listener);
-    listener({ type: "initial", entries: getRegistryEntries() });
-    return () => {
-        listeners.delete(listener);
-    };
-};
