+++ salt-marcher/tests/cartographer/AGENTS.md
@@ -1,9 +1,9 @@
 # Ziele
-- Prüft Cartographer-Modi, Presenter und Travel-spezifische Logik.
+- Prüft Cartographer-Modi, Controller und Travel-spezifische Logik.
 
 # Aktueller Stand
 - Unterordner `editor` und `travel` decken Werkzeuge sowie Reisecontroller ab.
-- Wurzeltests prüfen Presenter und Modusregistrierung.
+- Wurzeltests prüfen den Controller-Lifecycle samt Mode-Wechsel.
 
 # ToDo
 - keine offenen ToDos.
